+++ mdparser_arginfo.h
@@ -1,5 +1,5 @@
 /* This is a generated file, edit the .stub.php file instead.
- * Stub hash: f9a6f029f5258290e3e9d23c12587346ec4d387b */
+ * Stub hash: cdfe6bf9ce510c6a83b4e5d7a6dc3cce14269290 */
 
 ZEND_BEGIN_ARG_INFO_EX(arginfo_class_MdParser_Options___construct, 0, 0, 0)
 	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, sourcepos, _IS_BOOL, 0, "false")